app/requirements.txt
flask
opencv-python
numpy
kociemba
pillow
